+++ src/test/java/nextstep/subway/auth/TokenAuthenticationInterceptorTest.java
@@ -4,11 +4,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.subway.auth.application.UserDetailsService;
 import nextstep.subway.auth.domain.Authentication;
 import nextstep.subway.auth.domain.AuthenticationToken;
+import nextstep.subway.auth.domain.converter.AuthenticationConverter;
+import nextstep.subway.auth.domain.converter.TokenAuthenticationConverter;
 import nextstep.subway.auth.dto.TokenRequest;
 import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.auth.infrastructure.JwtTokenProvider;
-import nextstep.subway.auth.infrastructure.converter.AuthenticationConverter;
-import nextstep.subway.auth.infrastructure.converter.TokenAuthenticationConverter;
 import nextstep.subway.auth.ui.token.TokenAuthenticationInterceptor;
 import nextstep.subway.member.application.CustomUserDetailsService;
 import nextstep.subway.member.domain.LoginMember;
